we got notes
