+++ frontend/src/components/SavedRecipesIndex/SavedRecipeShow.jsx
@@ -1,24 +1,25 @@
 import React, { useState, useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { updateSavedRecipeNote } from '../../store/savedRecipes';
+import PostIt from './PostIt';
 import './SavedRecipeShow.css';
 
 const SavedRecipeShow = ({ recipe }) => {
-    const [note, setNote] = useState(recipe.note);
-    const dispatch = useDispatch();
+    // const [note, setNote] = useState(recipe.note);
+    // const dispatch = useDispatch();
 
-    useEffect(() => {
-        setNote(recipe.note);
-    }, [recipe]);
+    // useEffect(() => {
+    //     setNote(recipe.note);
+    // }, [recipe]);
 
     // console.log(note)
 
-    const handleSubmit = (e) => {
-        e.preventDefault();
-        debugger
-        dispatch(updateSavedRecipeNote(recipe._id, note));
-        
-    };
+    // const handleSubmit = (e) => {
+    //     e.preventDefault();
+    //     debugger
+    //     dispatch(updateSavedRecipeNote(recipe._id, note));
+
+    // };
 
     return (
         <div className='saved-recipe-show'>
@@ -35,7 +36,7 @@ const SavedRecipeShow = ({ recipe }) => {
                     <li key={index}>{instruction.step}. {instruction.description}</li>
                 ))}
             </ol>
-            <h3>Note</h3>
+            {/* <h3>Note</h3>
             <form onSubmit={handleSubmit}>
                 <textarea
                     className="note"
@@ -43,8 +44,9 @@ const SavedRecipeShow = ({ recipe }) => {
                     onChange={(e) => setNote(e.target.value)}
                 />
                 <button type="submit">Save Note</button>
-            </form>
-            <li>{recipe.note}</li>
+            </form> */}
+            <PostIt recipe={recipe} />
+            {/* <li>{recipe.note}</li> */}
         </div>
     );
 };
